requirements.txt
appdirs==1.4.4
Babel==2.9.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
Brotli==1.0.9
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
cssselect==1.1.0
cycler==0.10.0
dash==1.20.0
dash-bootstrap-components==0.12.0
dash-core-components==1.16.0
dash-html-components==1.1.3
dash-renderer==1.9.1
dash-table==4.11.3
DateTime==4.3
et-xmlfile==1.1.0
fake-useragent==0.1.11
feedparser==6.0.2
Flask==1.1.2
Flask-Caching==1.10.1
Flask-Compress==1.9.0
forex-python==1.5
future==0.18.2
greenlet==1.0.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
kiwisolver==1.3.1
lxml==4.6.3
MarkupSafe==1.1.1
matplotlib==3.4.1
numpy==1.20.2
openpyxl==3.0.7
pandas==1.2.4
pandas-datareader==0.9.0
paramiko==2.7.2
parse==1.19.0
patsy==0.5.1
Pillow==8.2.0
plotly==4.14.3
psycopg2==2.8.6
pycparser==2.20
pyee==8.1.0
pykalman==0.9.5
PyNaCl==1.4.0
pyparsing==2.4.7
pyppeteer==0.2.5
pyquery==1.4.3
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
requests==2.25.1
requests-html==0.10.0
retrying==1.3.3
scikit-learn==0.24.2
scipy==1.6.3
seaborn==0.11.1
sgmllib3k==1.0.0
simplejson==3.17.2
six==1.15.0
sklearn==0.0
soupsieve==2.2.1
SQLAlchemy==1.4.11
sshtunnel==0.4.0
statsmodels==0.12.2
threadpoolctl==2.1.0
tqdm==4.60.0
trafalgar.py==0.1
urllib3==1.26.4
w3lib==1.22.0
websockets==8.1
Werkzeug==1.0.1
yahoo-fin==0.8.8
zope.interface==5.4.0
